Fix JWT expiration field type and update token validation logic
- Change JWT expiration field from usize to i64 in Claims struct
- Update JWT expiration validation to use i64 directly without casting
- Update token generation script to use i64 for expiration timestamps
- Refresh test tokens in tokens.json with updated expiration format
- Add/update research issues in PROGRESS.md for future work
- Add research validator prompt file

+++ mqtt-auth-biscuit/benchmarks/gen_tokens.rs
@@ -11,7 +11,7 @@ use std::time::{SystemTime, UNIX_EPOCH};
 #[derive(Debug, Serialize, Deserialize)]
 struct Claims {
     sub: String,
-    exp: usize,
+    exp: i64,
     roles: Option<Vec<String>>,
     pad: Option<String>,
 }
@@ -53,7 +53,7 @@ fn main() {
         let now = SystemTime::now()
             .duration_since(UNIX_EPOCH)
             .unwrap()
-            .as_secs() as usize;
+            .as_secs() as i64;
         let claims = Claims {
             sub: "client_1".to_string(),
             exp: now + 5,
